Move fetchTeam to @helpers/fetchTeam
