+++ packages/app-project/src/helpers/fetchTeam/index.js
@@ -1,0 +1,1 @@
+export { default } from './fetchTeam'
